Moving page_token to api package
