+++ api/assessment/metric.proto
@@ -118,8 +118,3 @@ message MetricImplementation {
   // The actual implementation
   string code = 3;
 }
-
-message PageToken {
-  int64 start = 1;
-  int32 size = 2;
-}
