+++ service/assessment/assessment.go
@@ -27,7 +27,6 @@ package assessment
 
 import (
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -41,6 +40,7 @@ import (
 	"clouditor.io/clouditor/api/evidence"
 	"clouditor.io/clouditor/api/orchestrator"
 	"clouditor.io/clouditor/policies"
+	"clouditor.io/clouditor/service"
 	service_orchestrator "clouditor.io/clouditor/service/orchestrator"
 
 	"github.com/google/uuid"
@@ -50,7 +50,6 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/structpb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
@@ -399,7 +398,7 @@ func (s *Service) ListAssessmentResults(_ context.Context, req *assessment.ListA
 
 	res = new(assessment.ListAssessmentResultsResponse)
 
-	start, end, res.NextPageToken, err = paginate(req, max)
+	start, end, res.NextPageToken, err = service.Paginate(req, max)
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "could not decode page token: %v", err)
 	}
@@ -620,83 +619,4 @@ func (svc *Service) recvEventsLoop() {
 
 		_ = svc.pe.HandleMetricEvent(event)
 	}
-}
-
-func encodeToken(t *assessment.PageToken) (b64token string, err error) {
-	var b []byte
-
-	b, err = proto.Marshal(t)
-	if err != nil {
-		return "", fmt.Errorf("error while marshaling protobuf message: %w", err)
-	}
-
-	b64token = base64.URLEncoding.EncodeToString(b)
-	return
-}
-
-func decodeToken(b64token string) (t *assessment.PageToken, err error) {
-	var b []byte
-
-	b, err = base64.URLEncoding.DecodeString(b64token)
-	if err != nil {
-		return nil, fmt.Errorf("error while decoding base64 token: %w", err)
-	}
-
-	t = new(assessment.PageToken)
-
-	err = proto.Unmarshal(b, t)
-	if err != nil {
-		return nil, fmt.Errorf("error while unmarshalling protobuf message: %w", err)
-	}
-
-	return
-}
-
-type PaginatedRequest interface {
-	GetPageToken() string
-	GetPageSize() int32
-}
-
-// paginate is a helper function that helps to paginate list requests. It parses the necessary
-// informaton out if a paginated request, e.g. the page token and the desired page size and returns
-// the offset values for the requested page as well as the next page token.
-func paginate(req PaginatedRequest, max int64) (start int64, end int64, nbt string, err error) {
-	var token *assessment.PageToken
-
-	if req.GetPageToken() == "" {
-		// We need a new page token
-		token = &assessment.PageToken{
-			Start: 0,
-			Size:  req.GetPageSize(),
-		}
-	} else {
-		// Try to decode our existing token
-		token, err = decodeToken(req.GetPageToken())
-		if err != nil {
-			return 0, 0, "", fmt.Errorf("could not decode page token: %w", err)
-		}
-	}
-
-	start = token.Start
-	end = token.Start + int64(token.Size)
-	if end >= max {
-		end = max
-
-		// Indicate that we are at the end
-		token = nil
-	}
-
-	// Only needed, if more pages exist
-	if token != nil {
-		// Move the token "forward"
-		token.Start = end
-
-		// Encode next page token
-		nbt, err = encodeToken(token)
-		if err != nil {
-			return 0, 0, "", status.Errorf(codes.Internal, "could not create page token: %v", err)
-		}
-	}
-
-	return
 }
